Created new unreal-engine plugin folder
It's a pretty messy set here but located in the public folder myclass.h is the library file for the scan-plugin and the c++ file for it is myclass.cpp in the private folder.

All of the library functions for all things considering FBX are here, I had a hay-day trying to frankeinstein something together but I have not been able to get it working.