Pushing the Unreal demo in here for launching convenience
The main.py UI is able to launch this when you press the button, however the file pathway will need to be changed to match the user. If this does not work as intended the line on line 91 needs to be changed to where you have  locally downloaded the unreal executable.

This method works and does have the demo being launched from a button press in our python ui!